Mars web scraping project.
This is the result of using a combination of python, splinter, flask, beautiful soup, html and css to scrape data from several web services in order to display all of it on one page.  Unfortunately, it seems as though github does not support python/flask on it's "Pages" service. If it had, I would have hosted this site there. However, the attached "Mars_Webscrape.htm" gives an overview of what the results webpage would look like after scraping the data.